requirements.txt
asgiref==3.3.1
blessings==1.7
coverage==5.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-nose==1.4.7
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==5.1.0
importlib-metadata==3.3.0
Markdown==3.3.3
mock==4.0.3
nose==1.3.7
nose-progressive==1.5.2
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
zipp==3.4.0
